Move `PackageKind` to separate module
